+++ packages/ub-pub/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.11]
+### Fixed
+- for environment with SystemJS (usually browser) package will expose
+ itself and `@unitybase/cryptojs` as SystemJS module to prevent double-loading
+- injection.addResourceVersion correctly add version to URI with parameters (if `?` exists will use `&` as separator)
+
 ## [5.0.10]
 ### Fixed
 - request for retrieve data for cached entity from server will remove `logicalPredicates`